+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [runtime] WLAN network to use 192.168.144.0/24 network by default instead of 192.168.2.0/24
 - [runtime] WLAN DHCP to use 192.168.144.128/25 network by default instead of 192.168.2.128/25
+- [runtime] WiFi AP stack manually started in `fromfile` after container stack is reported Up.
 - [captive-portal] Using 1.6.0 which default 192.168.144.128/25 network by default instead of 192.168.2.128/25
 - [builder] Using 192.168.144.0/24 and 192.168.144.128/25 instead of 192.168.2.0/24 and 192.168.2.128/25
 - [builder] Added optional `variant` param