+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.7.0]
+## [Unreleased]
+
+### Changed
+
+- [runtime] `ap` reloads sysctl via `--system` instead of `-p` which fails on newer versions
+
+## [2.7.0] - 2025-09-04
 
 ### Changed